Fix a crash issue when close wxwebview_chromium sample.
On Mac OS X, it will trigger a crash issue when user close the window
(It also appear in other platforms, but the crash doesn't appear.)

The crash is caused by calling `CefShutdown` in wxApp::OnExit(). When
wx application running in wxApp::OnExit function, the webframe has been
deleted, we need to call `CefShutdown` in webframe `Close` period.